+++ x-pack/plugin/inference/src/main/java/org/elasticsearch/xpack/inference/external/openai/OpenAiEmbeddingsResponseEntity.java
@@ -3,11 +3,9 @@
  * or more contributor license agreements. Licensed under the Elastic License
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
- *
- * this file was contributed to by a generative AI
  */
 
-package org.elasticsearch.xpack.inference.external.response.openai;
+package org.elasticsearch.xpack.inference.external.openai;
 
 import org.elasticsearch.xcontent.ConstructingObjectParser;
 import org.elasticsearch.xcontent.ParseField;
