+++ x-pack/plugin/inference/src/main/java/org/elasticsearch/xpack/inference/external/openai/OpenAiRequest.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.external.request.openai;
+package org.elasticsearch.xpack.inference.external.openai;
 
 import org.elasticsearch.xpack.inference.external.request.Request;
